--- a/src/System.Memory/src/System/Buffers/ReadOnlySequence.cs
+++ b/src/System.Memory/src/System/Buffers/ReadOnlySequence.cs
@@ -37,7 +37,11 @@ namespace System.Buffers
         /// <summary>
         /// Determines if the <see cref="ReadOnlySequence{T}"/> contains a single <see cref="ReadOnlyMemory{T}"/> segment.
         /// </summary>
-        public bool IsSingleSegment => _sequenceStart.GetObject() == _sequenceEnd.GetObject();
+        public bool IsSingleSegment
+        {
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            get => _sequenceStart.GetObject() == _sequenceEnd.GetObject();
+        }
 
         /// <summary>
         /// Gets <see cref="ReadOnlyMemory{T}"/> from the first segment.
@@ -76,7 +80,7 @@ namespace System.Buffers
         /// Creates an instance of <see cref="ReadOnlySequence{T}"/> from linked memory list represented by start and end segments
         /// and corresponding indexes in them.
         /// </summary>
-        public ReadOnlySequence(IMemoryList<T> startSegment, int startIndex, IMemoryList<T> endSegment, int endIndex)
+        public ReadOnlySequence(ReadOnlySequenceSegment<T> startSegment, int startIndex, ReadOnlySequenceSegment<T> endSegment, int endIndex)
         {
             if (startSegment == null ||
                 endSegment == null ||
@@ -85,8 +89,8 @@ namespace System.Buffers
                 (startSegment == endSegment && endIndex < startIndex))
                 ThrowHelper.ThrowArgumentValidationException(startSegment, startIndex, endSegment);
 
-            _sequenceStart = new SequencePosition(startSegment, ReadOnlySequence.MemoryListToSequenceStart(startIndex));
-            _sequenceEnd = new SequencePosition(endSegment, ReadOnlySequence.MemoryListToSequenceEnd(endIndex));
+            _sequenceStart = new SequencePosition(startSegment, ReadOnlySequence.SegmentToSequenceStart(startIndex));
+            _sequenceEnd = new SequencePosition(endSegment, ReadOnlySequence.SegmentToSequenceEnd(endIndex));
         }
 
         /// <summary>
@@ -394,7 +398,7 @@ namespace System.Buffers
 
         private enum SequenceType
         {
-            IMemoryList = 0x00,
+            Segment = 0x00,
             Array = 0x1,
             OwnedMemory = 0x2,
             String = 0x3,
@@ -407,8 +411,8 @@ namespace System.Buffers
         public const int FlagBitMask = 1 << 31;
         public const int IndexBitMask = ~FlagBitMask;
 
-        public const int MemoryListStartMask = 0;
-        public const int MemoryListEndMask = 0;
+        public const int SegmentStartMask = 0;
+        public const int SegmentEndMask = 0;
 
         public const int ArrayStartMask = 0;
         public const int ArrayEndMask = FlagBitMask;
@@ -420,9 +424,9 @@ namespace System.Buffers
         public const int StringEndMask = FlagBitMask;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int MemoryListToSequenceStart(int startIndex) => startIndex | MemoryListStartMask;
+        public static int SegmentToSequenceStart(int startIndex) => startIndex | SegmentStartMask;
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int MemoryListToSequenceEnd(int endIndex) => endIndex | MemoryListEndMask;
+        public static int SegmentToSequenceEnd(int endIndex) => endIndex | SegmentEndMask;
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int ArrayToSequenceStart(int startIndex) => startIndex | ArrayStartMask;
         [MethodImpl(MethodImplOptions.AggressiveInlining)]