--- a/src/System.Memory/src/System/Buffers/ReadOnlySequenceSegment.cs
+++ b/src/System.Memory/src/System/Buffers/ReadOnlySequenceSegment.cs
@@ -5,7 +5,7 @@
 namespace System.Buffers
 {
     /// <summary>
-    /// Represents a linked list of <see cref="Memory{T}"/> nodes.
+    /// Represents a linked list of <see cref="ReadOnlyMemory{T}"/> nodes.
     /// </summary>
     public abstract class ReadOnlySequenceSegment<T>
     {