… src/System.Memory/src/System/Runtime/InteropServices/SequenceMarshal.cs
… src/System.Memory/src/System/ThrowHelper.cs
… src/System.Memory/tests/ReadOnlyBuffer/BufferSegment.cs
… src/System.Memory/tests/ReadOnlyBuffer/ReadOnlySequenceTests.Common.cs