Change `IMemoryList` to POCO object; then inherit for Pipeline`s `BufferSegment`; this means all its properties inline in use rather than being interface dispatch.

Improve some of the code paths
    
Bottom numbers after the effect for 1, 100 and 1000 segments; top is current

```        
                           Method |    Categories |        Mean |         Op/s | Scaled |
--------------------------------- |-------------- |------------:|-------------:|-------:|
   ReadOnlySequence<T> (previous) |     1 segment |   104.86 ns |  9,536,378.9 |   1.00 |
    ReadOnlySequence<T> (current) |     1 segment |    85.64 ns | 11,677,040.0 |   0.82 |
    ReadOnlySequence<T> (this PR) |     1 segment |    70.79 ns | 14,125,877.7 |   0.68 |
                                  |               |             |              |        |
   ReadOnlySequence<T> (previous) |     2 segment |   188.92 ns |  5,293,250.6 |   1.00 |
    ReadOnlySequence<T> (current) |     2 segment |   181.64 ns |  5,505,269.8 |   0.96 |
    ReadOnlySequence<T> (this PR) |     2 segment |   111.34 ns |  8,981,178.5 |   0.59 |
                                  |               |             |              |        |
   ReadOnlySequence<T> (previous) |  100 segments | 1,297.69 ns |    770,599.7 |   1.00 |
    ReadOnlySequence<T> (current) |  100 segments |   966.50 ns |  1,034,659.7 |   0.74 |
    ReadOnlySequence<T> (this PR) |  100 segments |   233.30 ns |  4,286,286.4 |   0.18 |
                                  |               |             |              |        |
   ReadOnlySequence<T> (previous) | 1000 segments | 1,390.97 ns |    718,923.3 |   1.00 |
    ReadOnlySequence<T> (current) | 1000 segments | 1,095.54 ns |    912,790.8 |   0.79 |
    ReadOnlySequence<T> (this PR) | 1000 segments |   275.27 ns |  3,632,776.9 |   0.20 |
                                  |               |             |              |        |
                          Span<T> |       MM item |   147.34 ns |  6,787,204.7 |   0.54 |
                   BufferSlice<T> |       MM item |   152.70 ns |  6,548,884.8 |   0.56 |
   ReadOnlySequence<T> (previous) |       MM item |   272.44 ns |  3,670,500.0 |   1.00 |
    ReadOnlySequence<T> (current) |       MM item |   251.15 ns |  3,981,724.7 |   0.92 |
    ReadOnlySequence<T> (this PR) |       MM item |   200.79 ns |  4,980,437.7 |   0.74 |
```

Resolves https://github.com/dotnet/corefx/issues/27500