@benaadams 2018-02-27
```       
                                                  Method |       Mean |        Op/s | Scaled | Allocated |
-------------------------------------------------------- |-----------:|------------:|-------:|----------:|
                'MM item. ReadOnlySequence<T> (current)' | 272.563 ns | 3,668,870.8 |   1.84 |       0 B |
 'MM item. ReadOnlySequence<T> (PR dotnet/corefx#27455)' | 254.244 ns | 3,933,232.7 |   1.71 |       0 B |
 'MM item. ReadOnlySequence<T> (PR dotnet/corefx#27499)' | 211.564 ns | 4,726,706.1 |   1.43 |       0 B |
```