@benaadams 2018-02-27
```
 System.Net.Http.Functional.Tests.HttpRetryProtocolTests/PostAsyncExpect100Continue_FailsAfterContentSendStarted_Throws
```
```
System.Exception : Unexpected EOF trying to read request header
   at System.Net.Test.Common.LoopbackServer.Connection.ReadRequestHeaderAsync() in D:\j\workspace\windows-TGrou---f8ac6754\src\Common\tests\System\Net\Http\LoopbackServer.cs:line 350
   at System.Net.Http.Functional.Tests.HttpRetryProtocolTests.<>c__DisplayClass4_3.<<PostAsyncExpect100Continue_FailsAfterContentSendStarted_Throws>b__3>d.MoveNext() in D:\j\workspace\windows-TGrou---f8ac6754\src\System.Net.Http\tests\FunctionalTests\HttpRetryProtocolTests.cs:line 101
--- End of stack trace from previous location where exception was thrown ---
   at System.Net.Test.Common.LoopbackServer.AcceptConnectionAsync(Func`2 funcAsync) in D:\j\workspace\windows-TGrou---f8ac6754\src\Common\tests\System\Net\Http\LoopbackServer.cs:line 117
   at System.Net.Http.Functional.Tests.HttpRetryProtocolTests.<>c__DisplayClass4_0.<<PostAsyncExpect100Continue_FailsAfterContentSendStarted_Throws>b__1>d.MoveNext() in D:\j\workspace\windows-TGrou---f8ac6754\src\System.Net.Http\tests\FunctionalTests\HttpRetryProtocolTests.cs:line 92
--- End of stack trace from previous location where exception was thrown ---
   at System.Net.Test.Common.LoopbackServer.<>c__DisplayClass11_0.<<CreateClientAndServerAsync>b__0>d.MoveNext() in D:\j\workspace\windows-TGrou---f8ac6754\src\Common\tests\System\Net\Http\LoopbackServer.cs:line 84
--- End of stack trace from previous location where exception was thrown ---
   at System.Net.Test.Common.LoopbackServer.CreateServerAsync(Func`2 funcAsync, Options options) in D:\j\workspace\windows-TGrou---f8ac6754\src\Common\tests\System\Net\Http\LoopbackServer.cs:line 67
   at System.Net.Http.Functional.Tests.HttpRetryProtocolTests.PostAsyncExpect100Continue_FailsAfterContentSendStarted_Throws() in D:\j\workspace\windows-TGrou---f8ac6754\src\System.Net.Http\tests\FunctionalTests\HttpRetryProtocolTests.cs:line 70
--- End of stack trace from previous location where exception was thrown ---
```