@benaadams 2018-02-27
@dotnet-bot test Windows x86 Release Build