@benaadams 2018-02-28
Am I only merge clashing with my previous commits? Want to know if can go for a high speed rebase or a careful pay attention rebase; git seems to have lots of conflicts; whereas shouldn't :-/