@benaadams 2018-02-28
Think am close; just checking asm output